requirements.txt
fastapi==0.95.2
pydantic==1.10.8
uvicorn==0.22.0
pandas==1.5.3
pyreadr==0.4.7
numpy==1.24.2
pyyaml
prophet==1.1.4
scipy==1.10.1
sidetable==0.9.1
httpx==0.24.1
statsmodels==0.14.0
pyarrow==13.0.0